+++ src/so_packages.sh
@@ -2,11 +2,12 @@
 
 # A list of packages to install via apt separated by a single space
 # You may use 'apt-cache search <package-name>' if you wanto to search packages
-SO_PACKAGES="git lm-sensors jq httpie gnome-tweaks steam-installer tmux discord virtualbox ksnip xclip wine solaar sqlite3"
+BASE_PACKAGES="curl wget git build-essential unzip ca-certificates gnupg"
+SO_PACKAGES="lm-sensors jq steam-installer tmux discord xclip solaar sqlite3"
 
 # Install basic apt and flatpak packages
-sudo apt install -y $SO_PACKAGES
 sudo apt update -y
+sudo apt install -y $SO_PACKAGES
 
 # Clean up
 sudo apt autoremove -y